+++ tt_metal/impl/dispatch/command_queue_common.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -11,17 +11,17 @@ enum class CoreType;
 
 namespace tt::tt_metal {
 
-uint32_t get_relative_cq_offset(uint8_t cq_id, uint32_t cq_size) { return cq_id * cq_size; }
+size_t get_relative_cq_offset(uint8_t cq_id, size_t cq_size) { return cq_id * cq_size; }
 
 uint16_t get_umd_channel(uint16_t channel) { return channel & 0x3; }
 
-uint32_t get_absolute_cq_offset(uint16_t channel, uint8_t cq_id, uint32_t cq_size) {
+size_t get_absolute_cq_offset(uint16_t channel, uint8_t cq_id, size_t cq_size) {
     return (DispatchSettings::MAX_HUGEPAGE_SIZE * get_umd_channel(channel)) +
            ((channel >> 2) * DispatchSettings::MAX_DEV_CHANNEL_SIZE) + get_relative_cq_offset(cq_id, cq_size);
 }
 
 template <bool addr_16B>
-uint32_t get_cq_issue_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
+size_t get_cq_issue_rd_ptr(chip_id_t chip_id, uint8_t cq_id, size_t cq_size) {
     uint32_t recv;
     chip_id_t mmio_device_id = tt::tt_metal::MetalContext::instance().get_cluster().get_associated_mmio_device(chip_id);
     uint16_t channel = tt::tt_metal::MetalContext::instance().get_cluster().get_assigned_channel_for_device(chip_id);
@@ -40,11 +40,11 @@ uint32_t get_cq_issue_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size)
     return recv;
 }
 
-template uint32_t get_cq_issue_rd_ptr<true>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
-template uint32_t get_cq_issue_rd_ptr<false>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
+template size_t get_cq_issue_rd_ptr<true>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
+template size_t get_cq_issue_rd_ptr<false>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
 
 template <bool addr_16B>
-uint32_t get_cq_issue_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
+size_t get_cq_issue_wr_ptr(chip_id_t chip_id, uint8_t cq_id, size_t cq_size) {
     uint32_t recv;
     chip_id_t mmio_device_id = tt::tt_metal::MetalContext::instance().get_cluster().get_associated_mmio_device(chip_id);
     uint16_t channel = tt::tt_metal::MetalContext::instance().get_cluster().get_assigned_channel_for_device(chip_id);
@@ -58,11 +58,11 @@ uint32_t get_cq_issue_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size)
     return recv;
 }
 
-template uint32_t get_cq_issue_wr_ptr<true>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
-template uint32_t get_cq_issue_wr_ptr<false>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
+template size_t get_cq_issue_wr_ptr<true>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
+template size_t get_cq_issue_wr_ptr<false>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
 
 template <bool addr_16B>
-uint32_t get_cq_completion_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
+size_t get_cq_completion_wr_ptr(chip_id_t chip_id, uint8_t cq_id, size_t cq_size) {
     uint32_t recv;
     chip_id_t mmio_device_id = tt::tt_metal::MetalContext::instance().get_cluster().get_associated_mmio_device(chip_id);
     uint16_t channel = tt::tt_metal::MetalContext::instance().get_cluster().get_assigned_channel_for_device(chip_id);
@@ -81,11 +81,11 @@ uint32_t get_cq_completion_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_
     return recv;
 }
 
-template uint32_t get_cq_completion_wr_ptr<true>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
-template uint32_t get_cq_completion_wr_ptr<false>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
+template size_t get_cq_completion_wr_ptr<true>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
+template size_t get_cq_completion_wr_ptr<false>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
 
 template <bool addr_16B>
-inline uint32_t get_cq_completion_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
+inline size_t get_cq_completion_rd_ptr(chip_id_t chip_id, uint8_t cq_id, size_t cq_size) {
     uint32_t recv;
     chip_id_t mmio_device_id = tt::tt_metal::MetalContext::instance().get_cluster().get_associated_mmio_device(chip_id);
     uint16_t channel = tt::tt_metal::MetalContext::instance().get_cluster().get_assigned_channel_for_device(chip_id);
@@ -99,7 +99,7 @@ inline uint32_t get_cq_completion_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint3
     return recv;
 }
 
-template uint32_t get_cq_completion_rd_ptr<true>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
-template uint32_t get_cq_completion_rd_ptr<false>(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
+template size_t get_cq_completion_rd_ptr<true>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
+template size_t get_cq_completion_rd_ptr<false>(chip_id_t chip_id, uint8_t cq_id, size_t cq_size);
 
 }  // namespace tt::tt_metal
